separate testData logic into a hook
